// packages/jupyter-pieces/jupyter_pieces-1.0.0.tar.gz/jupyter_pieces-1.0.0/src/PiecesSDK/core/models/Asset.ts
export interface Asset {
    /**
     *
     * @type {EmbeddedModelSchema}
     * @memberof Asset
     */
    schema?: EmbeddedModelSchema;
    /**
     * The globally available UID representing the asset in the Database, both locally and in the cloud.
     * @type {string}
     * @memberof Asset
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof Asset
     */
    name?: string;
    /**
     *
     * @type {string}
     * @memberof Asset
     */
    description?: string;
    /**
     *
     * @type {string}
     * @memberof Asset
     */
    creator: string;
    /**
     *
     * @type {GroupedTimestamp}
     * @memberof Asset
     */
    created: GroupedTimestamp;
    /**
     *
     * @type {GroupedTimestamp}
     * @memberof Asset
     */
    updated: GroupedTimestamp;
    /**
     *
     * @type {GroupedTimestamp}
     * @memberof Asset
     */
    synced?: GroupedTimestamp;
    /**
     *
     * @type {GroupedTimestamp}
     * @memberof Asset
     */
    deleted?: GroupedTimestamp;
    /**
     *
     * @type {AssetFormats}
     * @memberof Asset
     */
    formats: AssetFormats;
    /**
     *
     * @type {Preview}
     * @memberof Asset
     */
    preview: Preview;
    /**
     *
     * @type {ReferencedFormat}
     * @memberof Asset
     */
    original: ReferencedFormat;
    /**
     *
     * @type {Shares}
     * @memberof Asset
     */
    shares?: Shares;
    /**
     *
     * @type {MechanismEnum}
     * @memberof Asset
     */
    mechanism: MechanismEnum;
    /**
     *
     * @type {Websites}
     * @memberof Asset
     */
    websites?: Websites;
    /**
     *
     * @type {GroupedTimestamp}
     * @memberof Asset
     */
    interacted?: GroupedTimestamp;
    /**
     *
     * @type {Tags}
     * @memberof Asset
     */
    tags?: Tags;
    /**
     *
     * @type {Sensitives}
     * @memberof Asset
     */
    sensitives?: Sensitives;
    /**
     *
     * @type {Persons}
     * @memberof Asset
     */
    persons?: Persons;
    /**
     * This is an optional boolean that will flag that this asset came from a currated collection.
     * @type {boolean}
     * @memberof Asset
     */
    curated?: boolean;
    /**
     *
     * @type {boolean}
     * @memberof Asset
     */
    discovered?: boolean;
    /**
     *
     * @type {Activities}
     * @memberof Asset
     */
    activities?: Activities;
    /**
     *
     * @type {Score}
     * @memberof Asset
     */
    score?: Score;
    /**
     *
     * @type {boolean}
     * @memberof Asset
     */
    favorited?: boolean;
}